chore: more deprecated (re-)exports to `deprecated.ts`
