+++ plugins/catalog-backend/src/deprecated.ts
@@ -15,6 +15,9 @@
  */
 
 import {
+  type AnalyzeLocationEntityField as _AnalyzeLocationEntityField,
+  type AnalyzeLocationExistingEntity as _AnalyzeLocationExistingEntity,
+  type AnalyzeLocationGenerateEntity as _AnalyzeLocationGenerateEntity,
   type AnalyzeLocationRequest as _AnalyzeLocationRequest,
   type AnalyzeLocationResponse as _AnalyzeLocationResponse,
   type LocationSpec as _LocationSpec,
@@ -47,6 +50,12 @@ import {
   type PlaceholderResolverResolveUrl as _PlaceholderResolverResolveUrl,
   type ScmLocationAnalyzer as _ScmLocationAnalyzer,
 } from '@backstage/plugin-catalog-node';
+import {
+  defaultCatalogCollatorEntityTransformer as _defaultCatalogCollatorEntityTransformer,
+  DefaultCatalogCollatorFactory as _DefaultCatalogCollatorFactory,
+  type CatalogCollatorEntityTransformer as _CatalogCollatorEntityTransformer,
+  type DefaultCatalogCollatorFactoryOptions as _DefaultCatalogCollatorFactoryOptions,
+} from '@backstage/plugin-search-backend-module-catalog';
 
 /**
  * @public
@@ -208,3 +217,60 @@ export type AnalyzeLocationRequest = _AnalyzeLocationRequest;
  * @deprecated use the same type from `@backstage/plugin-catalog-common` instead
  */
 export type AnalyzeLocationResponse = _AnalyzeLocationResponse;
+
+/**
+ * If the folder pointed to already contained catalog info yaml files, they are
+ * read and emitted like this so that the frontend can inform the user that it
+ * located them and can make sure to register them as well if they weren't
+ * already
+ * @public
+ * @deprecated use the same type from `@backstage/plugin-catalog-common` instead
+ */
+export type AnalyzeLocationExistingEntity = _AnalyzeLocationExistingEntity;
+/**
+ * This is some form of representation of what the analyzer could deduce.
+ * We should probably have a chat about how this can best be conveyed to
+ * the frontend. It'll probably contain a (possibly incomplete) entity, plus
+ * enough info for the frontend to know what form data to show to the user
+ * for overriding/completing the info.
+ * @public
+ * @deprecated use the same type from `@backstage/plugin-catalog-common` instead
+ */
+export type AnalyzeLocationGenerateEntity = _AnalyzeLocationGenerateEntity;
+
+/**
+ *
+ * This is where I get really vague. Something like this perhaps? Or it could be
+ * something like a json-schema that contains enough info for the frontend to
+ * be able to present a form and explanations
+ * @public
+ * @deprecated use the same type from `@backstage/plugin-catalog-common` instead
+ */
+export type AnalyzeLocationEntityField = _AnalyzeLocationEntityField;
+
+/**
+ * @public
+ * @deprecated import from `@backstage/plugin-search-backend-module-catalog` instead
+ */
+export const DefaultCatalogCollatorFactory = _DefaultCatalogCollatorFactory;
+
+/**
+ * @public
+ * @deprecated import from `@backstage/plugin-search-backend-module-catalog` instead
+ */
+export const defaultCatalogCollatorEntityTransformer =
+  _defaultCatalogCollatorEntityTransformer;
+
+/**
+ * @public
+ * @deprecated import from `@backstage/plugin-search-backend-module-catalog` instead
+ */
+export type DefaultCatalogCollatorFactoryOptions =
+  _DefaultCatalogCollatorFactoryOptions;
+
+/**
+ * @public
+ * @deprecated import from `@backstage/plugin-search-backend-module-catalog` instead
+ */
+export type CatalogCollatorEntityTransformer =
+  _CatalogCollatorEntityTransformer;
