+++ plugins/catalog-backend/src/ingestion/types.ts
@@ -17,43 +17,7 @@
 import {
   AnalyzeLocationRequest,
   AnalyzeLocationResponse,
-  AnalyzeLocationExistingEntity as NonDeprecatedAnalyzeLocationExistingEntity,
-  AnalyzeLocationGenerateEntity as NonDeprecatedAnalyzeLocationGenerateEntity,
-  AnalyzeLocationEntityField as NonDeprecatedAnalyzeLocationEntityField,
 } from '@backstage/plugin-catalog-common';
-
-/**
- * If the folder pointed to already contained catalog info yaml files, they are
- * read and emitted like this so that the frontend can inform the user that it
- * located them and can make sure to register them as well if they weren't
- * already
- * @public
- * @deprecated use the same type from `@backstage/plugin-catalog-common` instead
- */
-export type AnalyzeLocationExistingEntity =
-  NonDeprecatedAnalyzeLocationExistingEntity;
-/**
- * This is some form of representation of what the analyzer could deduce.
- * We should probably have a chat about how this can best be conveyed to
- * the frontend. It'll probably contain a (possibly incomplete) entity, plus
- * enough info for the frontend to know what form data to show to the user
- * for overriding/completing the info.
- * @public
- * @deprecated use the same type from `@backstage/plugin-catalog-common` instead
- */
-export type AnalyzeLocationGenerateEntity =
-  NonDeprecatedAnalyzeLocationGenerateEntity;
-
-/**
- *
- * This is where I get really vague. Something like this perhaps? Or it could be
- * something like a json-schema that contains enough info for the frontend to
- * be able to present a form and explanations
- * @public
- * @deprecated use the same type from `@backstage/plugin-catalog-common` instead
- */
-export type AnalyzeLocationEntityField =
-  NonDeprecatedAnalyzeLocationEntityField;
 
 /** @public */
 export type LocationAnalyzer = {
